Separate functions into commons file
